refsw: Sync to 10.07.2019 (AB3.2/18.2)
refsw 5878b4b9a09dce620dc395f89ddb5660dbb972d7

+++ magnum/portinginterface/vce/src/bvce.c
@@ -8470,7 +8470,7 @@ BVCE_Channel_Picture_Enqueue_isr(
                   /* Seed the next target PTS to equal the current STC */
                   BDBG_MSG(("Reset TargetPTS %#x --> %#x", uiTargetPTSin45Khz, uiCurrentSTCin45Khz));
                   uiTargetPTSin45Khz = uiCurrentSTCin45Khz;
-                  hVceCh->picture.stState.uiNextTargetPTSin360Khz = uiCurrentSTCin45Khz*8;
+                  hVceCh->picture.stState.uiNextTargetPTSin360Khz = (uint64_t)uiCurrentSTCin45Khz*8;
                   hVceCh->picture.stState.bNextTargetPTSin360KhzValid = true;
                }
